+++ packages/expo-apple-authentication/src/AppleAuthentication.ts
@@ -25,7 +25,7 @@ export async function signInAsync(
     throw new UnavailabilityError('expo-apple-authentication', 'signInAsync');
   }
   const requestOptions = {
-    ...(options || {}),
+    ...options,
     requestedOperation: AppleAuthenticationOperation.LOGIN,
   };
   const credential = await ExpoAppleAuthentication.requestAsync(requestOptions);